requirements.txt
beautifulsoup4==4.12.3
crewai==0.11.0
google-api-python-client==2.118.0
crewai
openai==1.12.0
pandas==2.2.1
python-dotenv==1.0.1
uvicorn
fastapi
pytz==2024.1
requests==2.31.0
tqdm==4.66.2
weaviate-client==4.5.0
yfinance==0.2.36
youtube-transcript-api==0.6.2
PyYAML==6.0.1